# apps/backend/openapi.json
{
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "MatchItemSchema": {
        "description": "自動マッチング結果 1 件分。",
        "properties": {
          "col": {
            "title": "Col",
            "type": "integer"
          },
          "currentValue": {
            "title": "Currentvalue",
            "type": "string"
          },
          "placeholder": {
            "title": "Placeholder",
            "type": "string"
          },
          "row": {
            "title": "Row",
            "type": "integer"
          }
        },
        "required": [
          "row",
          "col",
          "currentValue",
          "placeholder"
        ],
        "title": "MatchItemSchema",
        "type": "object"
      },
      "MatchScanRequest": {
        "description": "プレースホルダ自動マッチング要求。",
        "properties": {
          "data": {
            "items": {
              "items": {},
              "type": "array"
            },
            "title": "Data",
            "type": "array"
          },
          "mergeCells": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/MergeCellRangeSchema"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mergecells"
          },
          "sheetName": {
            "title": "Sheetname",
            "type": "string"
          },
          "strategy": {
            "default": "ordered",
            "enum": [
              "ordered",
              "key",
              "primary"
            ],
            "title": "Strategy",
            "type": "string"
          }
        },
        "required": [
          "sheetName",
          "data"
        ],
        "title": "MatchScanRequest",
        "type": "object"
      },
...
      "ReportContextRoot": {
        "additionalProperties": true,
        "description": "レポートコンテキストのルート。\nリレーションを Pydantic モデルのリストとして持ち、\n@computed_field を通じて旧辞書互換の ByRole や Ordered キーを展開する。",
        "properties": {
          "company": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/CompanyContext"
              },
              {
                "type": "null"
              }
            ]
          },
          "companyId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Companyid"
          },
          "controlNumber": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Controlnumber"
          },
          "createdAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Createdat"
          },
          "customData": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Customdata"
          },
          "id": {
            "format": "uuid",
            "title": "Id",
            "type": "string"
          },
          "reportClientPrimary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportClientItemContext"
              },
              {
                "type": "null"
              }
            ],
            "readOnly": true
          },
          "reportClientsByCompanyId": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ReportClientItemContext"
            },
            "readOnly": true,
            "title": "Reportclientsbycompanyid",
            "type": "object"
          },
          "reportClientsByRole": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ReportClientItemContext"
            },
            "readOnly": true,
            "title": "Reportclientsbyrole",
            "type": "object"
          },
          "reportClientsOrdered": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReportClientItemContext"
                },
                {
                  "type": "null"
                }
              ]
            },
            "readOnly": true,
            "title": "Reportclientsordered",
            "type": "array"
          },
          "reportOwnedInstrumentPrimary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportOwnedInstrumentItemContext"
              },
              {
                "type": "null"
              }
            ],
            "readOnly": true
          },
          "reportOwnedInstrumentsByOwnedInstrumentId": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ReportOwnedInstrumentItemContext"
            },
            "readOnly": true,
            "title": "Reportownedinstrumentsbyownedinstrumentid",
            "type": "object"
          },
          "reportOwnedInstrumentsByType": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/ReportOwnedInstrumentItemContext"
              },
              "type": "array"
            },
            "readOnly": true,
            "title": "Reportownedinstrumentsbytype",
            "type": "object"
          },
          "reportOwnedInstrumentsOrdered": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReportOwnedInstrumentItemContext"
                },
                {
                  "type": "null"
                }
              ]
            },
            "readOnly": true,
            "title": "Reportownedinstrumentsordered",
            "type": "array"
          },
          "reportSitePrimary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportSiteItemContext"
              },
              {
                "type": "null"
              }
            ],
            "readOnly": true
          },
          "reportSitesByRole": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ReportSiteItemContext"
            },
            "readOnly": true,
            "title": "Reportsitesbyrole",
            "type": "object"
          },
          "reportSitesOrdered": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReportSiteItemContext"
                },
                {
                  "type": "null"
                }
              ]
            },
            "readOnly": true,
            "title": "Reportsitesordered",
            "type": "array"
          },
          "reportSnapshot": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reportsnapshot"
          },
          "reportTitle": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporttitle"
          },
          "reportType": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reporttype"
          },
          "reportWorkerPrimary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/ReportWorkerItemContext"
              },
              {
                "type": "null"
              }
            ],
            "readOnly": true
          },
          "reportWorkersByRole": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ReportWorkerItemContext"
            },
            "readOnly": true,
            "title": "Reportworkersbyrole",
            "type": "object"
          },
          "reportWorkersByWorkerId": {
            "additionalProperties": {
              "$ref": "#/components/schemas/ReportWorkerItemContext"
            },
            "readOnly": true,
            "title": "Reportworkersbyworkerid",
            "type": "object"
          },
          "reportWorkersOrdered": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/ReportWorkerItemContext"
                },
                {
                  "type": "null"
                }
              ]
            },
            "readOnly": true,
            "title": "Reportworkersordered",
            "type": "array"
          },
          "schemaId": {
            "anyOf": [
              {
                "format": "uuid",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Schemaid"
          },
          "targetInstrumentPrimary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TargetInstrumentContext"
              },
              {
                "type": "null"
              }
            ],
            "readOnly": true
          },
          "targetInstrumentsById": {
            "additionalProperties": {
              "$ref": "#/components/schemas/TargetInstrumentContext"
            },
            "readOnly": true,
            "title": "Targetinstrumentsbyid",
            "type": "object"
          },
          "targetInstrumentsByTagNumber": {
            "additionalProperties": {
              "$ref": "#/components/schemas/TargetInstrumentContext"
            },
            "readOnly": true,
            "title": "Targetinstrumentsbytagnumber",
            "type": "object"
          },
          "targetInstrumentsOrdered": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/TargetInstrumentContext"
                },
                {
                  "type": "null"
                }
              ]
            },
            "readOnly": true,
            "title": "Targetinstrumentsordered",
            "type": "array"
          },
          "updatedAt": {
            "anyOf": [
              {
                "format": "date-time",
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Updatedat"
          },
          "usedPartPrimary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/UsedPartItemContext"
              },
              {
                "type": "null"
              }
            ],
            "readOnly": true
          },
          "usedPartsByCategory": {
            "additionalProperties": {
              "items": {
                "$ref": "#/components/schemas/UsedPartItemContext"
              },
              "type": "array"
            },
            "readOnly": true,
            "title": "Usedpartsbycategory",
            "type": "object"
          },
          "usedPartsById": {
            "additionalProperties": {
              "$ref": "#/components/schemas/UsedPartItemContext"
            },
            "readOnly": true,
            "title": "Usedpartsbyid",
            "type": "object"
          },
          "usedPartsOrdered": {
            "items": {
              "anyOf": [
                {
                  "$ref": "#/components/schemas/UsedPartItemContext"
                },
                {
                  "type": "null"
                }
              ]
            },
            "readOnly": true,
            "title": "Usedpartsordered",
            "type": "array"
          }
        },
        "required": [
          "id",
          "reportSitesByRole",
          "reportSitesOrdered",
          "reportSitePrimary",
          "reportClientsByCompanyId",
          "reportClientsByRole",
          "reportClientsOrdered",
          "reportClientPrimary",
          "reportWorkersByWorkerId",
          "reportWorkersByRole",
          "reportWorkersOrdered",
          "reportWorkerPrimary",
          "targetInstrumentsById",
          "targetInstrumentsByTagNumber",
          "targetInstrumentsOrdered",
          "targetInstrumentPrimary",
          "reportOwnedInstrumentsByOwnedInstrumentId",
          "reportOwnedInstrumentsByType",
          "reportOwnedInstrumentsOrdered",
          "reportOwnedInstrumentPrimary",
          "usedPartsById",
          "usedPartsByCategory",
          "usedPartsOrdered",
          "usedPartPrimary"
        ],
        "title": "ReportContextRoot",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      },
...
    }
  },
  "info": {
    "description": "現場報告管理システム Citadel のバックエンド API",
    "title": "Citadel API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/generate-excel": {
      "post": {
        "description": "report_id を受け取り、DB から関連データを取得し、Report に紐づくレポート種別\n(ReportFormat) に応じた複数テンプレートを個別 Excel として処理し、ZIP 圧縮して返す。",
        "operationId": "generate_excel_api_generate_excel_post",
        "parameters": [
          {
            "description": "出力対象のレポート ID",
            "in": "query",
            "name": "report_id",
            "required": true,
            "schema": {
              "description": "出力対象のレポート ID",
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Excel",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/generate-report": {
      "post": {
        "description": "report_id を受け取り、DB から関連データを取得し、Report に紐づくレポート種別\n(ReportFormat) に応じた複数テンプレートを順に処理し、1 つの PDF に結合して返す。",
        "operationId": "generate_report_api_generate_report_post",
        "parameters": [
          {
            "description": "出力対象のレポート ID",
            "in": "query",
            "name": "report_id",
            "required": true,
            "schema": {
              "description": "出力対象のレポート ID",
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          },
          {
            "description": "True の場合、Microsoft Print to PDF で印刷して PDF 出力（高画質）",
            "in": "query",
            "name": "use_printer",
            "required": false,
            "schema": {
              "default": false,
              "description": "True の場合、Microsoft Print to PDF で印刷して PDF 出力（高画質）",
              "title": "Use Printer",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Generate Report",
        "tags": [
          "reports"
        ]
      }
    },
...
    "/api/reports/{report_id}": {
      "delete": {
        "description": "指定IDのレポートを削除。関連データを子から親の順でカスケード削除する。",
        "operationId": "delete_report_api_reports__report_id__delete",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Delete Report",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/reports/{report_id}/complete": {
      "post": {
        "description": "レポートを「完了」にする。完了時点のコンテキストを report_snapshot に保存する。\n以降の PDF/Excel 出力ではスナップショットを使用するため、マスタ変更の影響を受けない。",
        "operationId": "complete_report_api_reports__report_id__complete_post",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {}
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Complete Report",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/reports/{report_id}/context": {
      "get": {
        "description": "指定レポートIDのデータをロードし、レポートコンテキストをそのまま返す。プレースホルダ生成ツール用。",
        "operationId": "get_report_context_api_reports__report_id__context_get",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReportContextRoot"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Get Report Context",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/reports/{report_id}/match-scan": {
      "post": {
        "description": "指定レポートのコンテキストとシートのセル値から、自動マッチング候補を返す。",
        "operationId": "run_report_match_scan_api_reports__report_id__match_scan_post",
        "parameters": [
          {
            "in": "path",
            "name": "report_id",
            "required": true,
            "schema": {
              "format": "uuid",
              "title": "Report Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/MatchScanRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/MatchItemSchema"
                  },
                  "title": "Response Run Report Match Scan Api Reports  Report Id  Match Scan Post",
                  "type": "array"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Run Report Match Scan",
        "tags": [
          "reports"
        ]
      }
    },
...
  }
}